Refactor : jwtUtil 위치 바꿈
